% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dbi.R
\name{dbClearResult,sparklyr2_result-method}
\alias{dbClearResult,sparklyr2_result-method}
\title{Clears sparklyr2 result}
\usage{
\S4method{dbClearResult}{sparklyr2_result}(res, ...)
}
\arguments{
\item{res}{An object inheriting from DBIResult.}

\item{...}{Other arguments passed on to methods.}
}
\description{
Clears sparklyr2 result
}
