% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dbi.R
\name{dbFetch,sparklyr2_result-method}
\alias{dbFetch,sparklyr2_result-method}
\title{Retrieve records from a Spark query}
\usage{
\S4method{dbFetch}{sparklyr2_result}(res, n = -1, ...)
}
\arguments{
\item{res}{An object inheriting from DBIResult, created by dbSendQuery().}

\item{n}{maximum number of records to retrieve per fetch. Use n = -1
or n = Inf to retrieve all pending records. Some implementations may
recognize other special values.}

\item{...}{Other arguments passed on to methods. description}
}
\description{
Retrieve records from a Spark query
}
