% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-methods.R
\name{dbi_driver}
\alias{dbi_driver}
\alias{dbConnect,Sparklyr2Driver-method}
\title{DBI integration}
\usage{
dbi_driver()

\S4method{dbConnect}{Sparklyr2Driver}(drv, host = "sc://localhost", ...)
}
\arguments{
\item{drv}{An object created by \code{Sparklyr2()}}
}
\description{
DBI integration
}
\examples{
\dontrun{
sc <- dbConnect(sparklyr2::dbi_driver())
dbWriteTable(sc, "mtcars", mtcars)
dbGetQuery(sc, "SELECT * FROM mtcars WHERE cyl == 4")
}
}
