% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dbi.R
\docType{class}
\name{DBIMethods}
\alias{DBIMethods}
\alias{dbSendQuery,sparklyr2_connection-method}
\alias{dbClearResult,sparklyr2_result-method}
\alias{dbFetch,sparklyr2_result-method}
\title{DBI connection and results methods}
\usage{
\S4method{dbSendQuery}{sparklyr2_connection}(conn, statement, ...)

\S4method{dbClearResult}{sparklyr2_result}(res, ...)

\S4method{dbFetch}{sparklyr2_result}(res, n = -1, ...)
}
\arguments{
\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{statement}{a character string containing SQL.}

\item{...}{Other parameters passed on to methods.}

\item{res}{An object inheriting from \linkS4class{DBIResult}.}

\item{n}{maximum number of records to retrieve per fetch. Use \code{n = -1}
or \code{n = Inf}
to retrieve all pending records.  Some implementations may recognize other
special values.}
}
\description{
DBI connection and results methods
}
