% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dbi.R
\name{dbSendQuery,sparklyr2_connection-method}
\alias{dbSendQuery,sparklyr2_connection-method}
\title{Send a query to Spark using sparklyr2}
\usage{
\S4method{dbSendQuery}{sparklyr2_connection}(conn, statement, ...)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by dbConnect().}

\item{statement}{A character string containing SQL.}

\item{...}{Other parameters passed on to methods.}
}
\description{
Send a query to Spark using sparklyr2
}
