% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{spark_disconnect}
\alias{spark_disconnect}
\title{Disconnects from a Spark environment}
\usage{
spark_disconnect(sc)
}
\arguments{
\item{sc}{Spark connection}
}
\description{
Disconnects from a Spark environment
}
