% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{spark_connect}
\alias{spark_connect}
\title{Connects to a Spark environment}
\usage{
spark_connect(host)
}
\arguments{
\item{host}{The URL or path for the master, or the remote, parameter}
}
\description{
Connects to a Spark environment
}
